navbar updates
