--- client/src/components/Header/header.scss
@@ -1,18 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Dancing+Script&display=swap");
-
-body {
-  font-family: cursive, monospace, sans-serif;
-}
-
-header {
-  h1 {
-    color: #4e8d7c;
-    text-shadow: -1px -1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000,
-      1px 1px 0 #000;
-    font-family: "Dancing Script", cursive;
-  }
-
-  .logo-image {
-    width: 150px;
-  }
-}
